+++ src/optimizers/optimizer.rs
@@ -4,5 +4,5 @@ use crate::core::layer::Layer;
 
 pub trait Optimizer {
     fn initialize_layer_additional_params(&self, layer: &mut Layer);
-    fn update_params(&mut self, batch_size: usize, layer: &mut Layer, learning_rate: f64);
+    fn update_params(&mut self, batch_size: usize, layer: &mut Layer, learning_rate: f32);
 }
